% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourToFreq.R
\name{contourToFreq}
\alias{contourToFreq}
\title{Add Frequency and Time to Pamguard Whistle Binaries}
\usage{
contourToFreq(data)
}
\arguments{
\item{data}{either a \code{PamBinary} class object or just the \code{$data} from
a PamBinary object}
}
\value{
\code{data} with items \code{freq} and \code{time} added. These use the 
  calculated FFT window length, hope size, and sample rate to compute the frequency
  and time values of the saved whistle contour
}
\description{
Adds items \code{freq} and \code{time} to a Pamguard binary file
  from the Whistle & Moan Detector
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
