% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millisToDateNum.R
\name{millisToDateNum}
\alias{millisToDateNum}
\title{Convert Java Millisecond Time to R}
\usage{
millisToDateNum(millis)
}
\arguments{
\item{millis}{Millisecond time from Java}
}
\value{
Numeric time used by R.
}
\description{
Converts Java millisecond time into numeric
  time that R uses.
}
\note{
Original function was more relevant as Matlab and Java
  use different time origins. Java & R both use 1970-01-01,
  but Java stores as milliseconds vs seconds in R.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
