% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbToDf.R
\name{pbToDf}
\alias{pbToDf}
\title{Convert a PamBinary Object to Data Frame}
\usage{
pbToDf(pb, templateNames = NULL)
}
\arguments{
\item{pb}{a PamBinary class object created by 
\code{\link[PamBinaries]{loadPamguardBinaryFile}}}

\item{templateNames}{if using the click template classifier, the names of the species
for the click templates. These will be used as the names of the columns in the
dataframe, and the length of this must exactly match the number of templates used}

\item{\dots}{Unused parameters passed in from other methods}
}
\value{
a data.frame containing most of the binary data read in. Will not
  contain annotation data, click waveforms, DIFAR demux data, or contour
  information from WMD detector. These are skipped because they are either
  incosistent in their size, or are large objects.
}
\description{
Converts a PamBinary object into a data frame. The data.frame
  will combine all of the data from the \code{data} part of the PamBinary
  object, but will not include annotations data, click waveforms, DIFAR demux
  data, or contours from the WMD detector. These are skipped because they are
  either inconsistent in their size, or are large objects. The function 
  \code{pbToDf} is also called when \code{as.data.frame} is called on a PamBinary
  class object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
