% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWMD.R
\name{plotWMD}
\alias{plotWMD}
\title{Plot Whistle Contour}
\usage{
plotWMD(data, id = 1)
}
\arguments{
\item{data}{either a \code{PamBinary} class object, or just the \code{$data} from
a PamBinary object, or a single detection from the \code{$data}}
}
\value{
Nothing, just plots
}
\description{
Plots the entire whistle contour saved in a Pamguard Whistle &
  Moan Detector binary file, highlighting the selected contour
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
