% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickBackground.R
\name{readClickBackground}
\alias{readClickBackground}
\title{Read Click Detector Background data}
\usage{
readClickBackground(fid, fileInfo, data)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header, module header, and the
appropriate function to read module specific data}

\item{data}{a structure containing standard data}
}
\value{
a structure containing data from a single object
}
\description{
Reads in the background data saved by the Click Detector
}
\author{
Michael Oswald \email{taiki.sakai@noaa.gov}
}
