% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickClsfrAnnotation.R
\name{readClickClsfrAnnotation}
\alias{readClickClsfrAnnotation}
\title{Read Click Classifier Annotation}
\usage{
readClickClsfrAnnotation(fid, fileInfo, debug = FALSE)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{debug}{logical flag to show more info on errors}
}
\value{
a vector of click classifiers, represented by the click type flag
}
\description{
Reads binary data stored by Click Classifier annotations
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
