% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickFooter.R
\name{readClickFooter}
\alias{readClickFooter}
\title{Read Click Footer}
\usage{
readClickFooter(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
footer information for Click Detector module
}
\description{
Reads module footer information for the Click Detector module.
  Note that sometimes there is no additional footer information, so check
  first whether or not the binaryLength variable is 0.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
