% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickTriggerHeader.R
\name{readClickTriggerHeader}
\alias{readClickTriggerHeader}
\title{Read Click Trigger Header}
\usage{
readClickTriggerHeader(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
header information for the click trigger
}
\description{
Reads file header information specific to the click trigger module
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
