% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDifarData.R
\name{readDifarData}
\alias{readDifarData}
\title{Read Difar Data}
\usage{
readDifarData(fid, fileInfo, data, skipLarge = FALSE, debug = FALSE)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{data}{a structure containing standard data}

\item{skipLarge}{a flag of whether or not to skip reading the waveform}

\item{debug}{logical flag to show more info on errors}
}
\value{
a structure containing data from a single object, and a logical
  flag if an error has occurred
}
\description{
Reads binary data stored by the Difar Processing module.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
