% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileFooterInfo.R
\name{readFileFooterInfo}
\alias{readFileFooterInfo}
\title{Read File Footer}
\usage{
readFileFooterInfo(fid, version)
}
\arguments{
\item{fid}{binary file to be read}

\item{version}{binary file version}
}
\value{
footer information common to all files
}
\description{
Reads in the binary file footer. The input variable version
  is the file format read in from the file header. As of version 3, the 
  file footer includes the lowest and highest UID values in the file.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
