% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileHeader.R
\name{readFileHeader}
\alias{readFileHeader}
\title{Read File Header}
\usage{
readFileHeader(file, readExtra = FALSE)
}
\arguments{
\item{file}{binary file to be read}

\item{readExtra}{flag if there is extra information to read}
}
\value{
header information common to all files
}
\description{
Reads file header information common to all files
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
