% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJavaUTFString.R
\name{readJavaUTFString}
\alias{readJavaUTFString}
\title{Read Java UTF-8 String}
\usage{
readJavaUTFString(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
the string and its length
}
\description{
Reads a Java UTF-8 string. The first 2 bytes are the
  length of the string, then the string itself.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
