% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLTSAHeader.R
\name{readLTSAHeader}
\alias{readLTSAHeader}
\title{Read LTSA Header}
\usage{
readLTSAHeader(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
header information for the LTSA module
}
\description{
Reads file header information specific to the LTSA module
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
