% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMatchClsfrAnnotation.R
\name{readMatchClsfrAnnotation}
\alias{readMatchClsfrAnnotation}
\title{Read Matched Classifier Annotation}
\usage{
readMatchClsfrAnnotation(fid, fileInfo, anVersion, debug = FALSE)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{anVersion}{version id of annotation module}

\item{debug}{logical flag to show more info on errors}
}
\value{
a vector with the threshold, matchcorr, and rejectcorr values. See description.
}
\description{
Reads annotations from the matched click classifier. The matched
   matched click classifier annotates click detections with a threshold, matchcorr
   and rejectcorr values. The threshold value is used in the binary classification
   process. If it exceeds a hard value then the click is classified with the
   set type. The matchcorr and rejectcorr values are simply the correlation 
   values of the match and reject templates with the click.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
