% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNoiseMonHeader.R
\name{readNoiseMonHeader}
\alias{readNoiseMonHeader}
\title{Read Noise Monitor Header}
\usage{
readNoiseMonHeader(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
header information for the Noise Monitor module
}
\description{
Reads file header information specific to the Noise Monitor module
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
