% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPamData.R
\name{readPamData}
\alias{readPamData}
\title{Read Pamguard Data}
\usage{
readPamData(fid, fileInfo, ...)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header, module header, and the
appropriate function to read module specific data}

\item{\dots}{Arguments passed to other functions}
}
\value{
a structure containing data from a single object
}
\description{
Reads in the object data that is common to all modules. This 
  reads up to (but not including) the object binary length, and then calls 
  a function to read the module-specific data.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
