% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpectralBackground.R
\name{readSpectralBackground}
\alias{readSpectralBackground}
\title{Read Spectral Background data}
\usage{
readSpectralBackground(fid, fileInfo, data)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header, module header, and the
appropriate function to read module specific data}

\item{data}{a structure containing standard data}
}
\value{
a structure containing data from a single object
}
\description{
Reads in the background data saved by various detectors (e.g
   WMD, Right Whale Edge Detector, etc) EXCEPT FOR the Click Detector
}
\author{
Michael Oswald \email{taiki.sakai@noaa.gov}
}
