% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStdModuleFooter.R
\name{readStdModuleFooter}
\alias{readStdModuleFooter}
\title{Read Standard Module Footer}
\usage{
readStdModuleFooter(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
footer information common to all modules
}
\description{
Reads the module footer information common to all modules.
  Differs from the legacy code in that it does not read in or skip any
  information specific to a module.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
