% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStdModuleHeader.R
\name{readStdModuleHeader}
\alias{readStdModuleHeader}
\title{Read Standard Module Header}
\usage{
readStdModuleHeader(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
header information common to all modules
}
\description{
Reads the module header information common to all modules.
  Differs from the legacy code in that it does not read in or skip any
  information specific to a module.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
