% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTDBLAnnotation.R
\name{readTDBLAnnotation}
\alias{readTDBLAnnotation}
\title{Read TDBL Annotation}
\usage{
readTDBLAnnotation(fid, fileInfo, anVersion, debug = FALSE, ...)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{anVersion}{annotation version}

\item{debug}{logical flag to show more info on errors}

\item{\dots}{Arguments passed to other functions}
}
\value{
a structure containing data from a single object, and a logical
  flag if an error has occurred
}
\description{
Reads binary data stored by TDBL annotation module
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
