% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWMDHeader.R
\name{readWMDHeader}
\alias{readWMDHeader}
\title{Read Whistle & Moan Detector Header}
\usage{
readWMDHeader(file)
}
\arguments{
\item{file}{binary file to be read}
}
\value{
header information for the Whistle & Moan Detector module
}
\description{
Reads file header information specific to the 
  Whistle & Moan Detector module
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
