% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickData.R
\name{readClickData}
\alias{readClickData}
\title{Read Click Data}
\usage{
readClickData(fid, fileInfo, data)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{data}{a structure containing standard data}
}
\value{
a structure containing data from a single object, and a logical
  flag if an error has occurred
}
\description{
Reads binary data stored by the Click Detector module.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
