% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClickData.R
\name{readClickData}
\alias{readClickData}
\title{Read Click Data}
\usage{
readClickData(
  fid,
  fileInfo,
  data,
  skipLarge = FALSE,
  debug = FALSE,
  getWave,
  onlyWave
)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{data}{a structure containing standard data}

\item{skipLarge}{a flag for whether or not to skip reading large wave file}

\item{debug}{logical flag to show more info on errors}

\item{getWave}{DEPRECATED: see skipLarge}

\item{onlyWave}{DEPRECATED: see skipLarge}
}
\value{
a structure containing data from a single object, and a logical
  flag if an error has occurred
}
\description{
Reads binary data stored by the Click Detector module.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
