% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWMDData.R
\name{readWMDData}
\alias{readWMDData}
\title{Read Whistle and Moan Data}
\usage{
readWMDData(fid, fileInfo, data, skipLarge = FALSE, debug = FALSE)
}
\arguments{
\item{fid}{binary file identifier}

\item{fileInfo}{structure holding the file header and module header}

\item{data}{a structure containing standard data}

\item{skipLarge}{a flag for whether or not to skip reading large contours}

\item{debug}{logical flag to show more info on errors}
}
\value{
a structure containing data from a single object, and a logical
  flag if an error has occurred
}
\description{
Reads binary data stored by the Whistle & Moan Detector
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
